ebdeed086ebeaf416912956ebf8178b1c0b005e3 Cleanup imports (sanket1729)
382c8f9e4f91bb06b41938c7d9d1acaa02d9f398 Introduce PsbtSigHashType (sanket1729)

Pull request description:

  We cannot really use `Psbt` for taproot because the sighash type is currently EcdsaSigHashType. We could introduce an enum with two options but then deser is not really clear, so I chose the approach in the current PR. Feedback or other ways to do this welcome :)

  This is NOT related to #776

ACKs for top commit:
  apoelstra:
    ACK ebdeed086ebeaf416912956ebf8178b1c0b005e3
  dr-orlovsky:
    ACK ebdeed086ebeaf416912956ebf8178b1c0b005e3

Tree-SHA512: f9424cf3db09098d73f0d431a45ff86a47f11f7d40785bf95e58991fd4d16f0db0a9a3a63f898628b29c95bbd2ca901312a6a44ac6d8aec73a6a34710f6354a2

+++ src/util/sighash.rs
@@ -167,7 +167,7 @@ pub enum Error {
     WrongAnnex,
 
     /// Invalid Sighash type
-    InvalidSigHashType(u8),
+    InvalidSigHashType(u32),
 }
 
 impl fmt::Display for Error {
@@ -286,7 +286,7 @@ impl SchnorrSigHashType {
             0x81 => Ok(SchnorrSigHashType::AllPlusAnyoneCanPay),
             0x82 => Ok(SchnorrSigHashType::NonePlusAnyoneCanPay),
             0x83 => Ok(SchnorrSigHashType::SinglePlusAnyoneCanPay),
-            x => Err(Error::InvalidSigHashType(x)),
+            x => Err(Error::InvalidSigHashType(x as u32)),
         }
     }
 }
